feat: make livewire/update universal
This implements changes to the blocking mechanism of the tenancy
library. This is done but creating a custom middleware for blocking
central domain access that allows request if the path is
`livewire/update`. This also implements a custon ::onFail function that
checks if the path is also `livewire/update` and if the check passes
then the request gets passed to the next middleware.